refactor: product register 페이지 리팩토링
